refactor: move study hook into model
