+++ src/features/study/model/useStudy.spec.tsx
